website: fix community page issues (#4904)
* Fix inconsistent spacing in community.css
* Fix background on community page so it doesn't become white when scrolling

The white background happened because it was forced to a height of the viewport,
but the overflowing child elements were still visible. Because its a parent
element that scrolls, the background would scroll with the page causing the
white background to appear.

The culprit is setting `height: 100%`, which means it won't pick up the height
from its child elements. If instead we set `min-height: 100%` it'll be at
minimum the height of its parent (`<body>`, which along with `<html>` is set to
be `100%` of the viewport width and height – this is probably not ideal) but it
can still expand to fit all its child elements.

There are other ways to fix this, but this is probably the minimally invasive.

* Make dividers no push beyond the viewport width, and more responsive

The dividers will no longer push beyond the viewport width with these changes
(key culprit: the `-1em` left/right margins) and also will disappear on smaller
screens so they don't become very thin.

Signed-off-by: Marcus Stade <marcus@styra.com>