internal/strvals: port helm strvals fix (CLI --set arguments)(#5315)
The impact of this in OPA is minimal: when passing for example

    --set "abc[9999999999999]=123"

to `opa run` or `opa exec`, it would run out of memory. With this change, it
will return an error instead:

    $ opa exec --set "abc[9999999999999]=123" .
    {"err":"config error: failed parsing --set data: index of 9999999999999 is greater than maximum supported index of 65536",
     "level":"error","msg":"Unexpected error.","time":"2022-10-27T10:24:01+02:00"}

Since these parameters are never user-controlled, or accessible at
runtime, any attack vector here involves having access to a machines
CLI, and is thus equivalent to already having local user rights. The
affected instance of OPA would be the instance that is started -- so no
others would be affected.

Thanks to @pjbgf for bringing this up on the security mailing list.

See also the issue in helm,
https://github.com/fluxcd/flux2/security/advisories/GHSA-p2g7-xwvr-rrw3.

The changes have been ported from @mattfarina fix in
https://github.com/helm/helm/commit/225f8d7732aba378378e7507655157b1d26cd514.

More stuff has happened on that package in helm, and new features have been
added -- but I've refrained from pulling all that in.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>