topdown+ast+rego: read lazy objects from the store (#5380)
Before, any call to a builtin that would reference an object from data would force that
the entire object was transformed to an ast.Value, and a copy of it was passed to the builtin.

Now, we're doing something a little more involved: when an object is read from data, we'll
wrap it into a special container and don't convert it. The conversion only happens when it's
required. As such, we can avoid quite a bit of memory usage, and avoiding the extra work,
make evaluations faster.

Lazy objects are also immutable -- calling Insert() on them will panic. But based on how
topdown works, it should never happen. Only Golang coding modifying the results of a query
evaluation could be affected by this. If you are, see the opt-out mechanisms outlined below.

⚠️ The benchmark that follows is an edge case, but the issue at hand here has come up in
the real world before.

    name                         old time/op    new time/op    delta
    MemberWithKeyFromBaseDoc-16    40.2ms ±13%     0.0ms ± 5%  -99.97%  (p=0.000 n=9+10)
    ObjectGetFromBaseDoc-16        42.4ms ± 4%     0.0ms ± 2%  -99.97%  (p=0.000 n=9+10)

    name                         old alloc/op   new alloc/op   delta
    MemberWithKeyFromBaseDoc-16    17.2MB ± 0%     0.0MB ± 0%  -99.96%  (p=0.000 n=10+10)
    ObjectGetFromBaseDoc-16        17.2MB ± 0%     0.0MB ± 0%  -99.95%  (p=0.000 n=10+10)

    name                         old allocs/op  new allocs/op  delta
    MemberWithKeyFromBaseDoc-16      702k ± 0%        0k ± 0%  -99.98%  (p=0.000 n=10+10)
    ObjectGetFromBaseDoc-16          702k ± 0%        0k ± 0%  -99.98%  (p=0.000 n=10+10)

Note: The optimization is enabled by default, and for all storage implementations. However,
there are multiple layers of opt-out options, in case this causes any unforeseen trouble for
your usage of OPA via Golang. (OPA-as-server use cases don't need to worry about this.)

Concretely, these opt-out options are:

1. an option for `topdown.Query` to not start with lazy objects in the first place
2. an option for `ast.JSONWithOpt` to revert to the old behaviour when converting to golang native values
3. an `EvalOption` for `rego.PrepareForEval(...).Eval(...)` so that it keeps copying the maps it adds to the result set

Fixes #5325.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>