ast: lazy obj follow-ups (#5394)
* ast/lazyObj.Keys(): return from strict if we've got it

Taking it from there should allow us to re-use the memory we're already using
with the strict *object. The Keys() function otherwise creates a new ast.Term
for each key.

* ast/lazyObj.Insert(): force on insert, don't panic

So with the panic() we've had there before, we've established that in our code,
the Insert() won't happen, so lazyObj can be treated as immutable.

However, it's conceivable that someone is using OPA-via-Golang, and calls
Insert on a lazyObj returned from a topdown query. While there are opt-out
options, it's still nicer satisfy the entire ast.Object interface if we can.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>