Use Normalized policy paths as compiler module keys and store ids
Typically the path of the policy is used as the policy ID when
inserting in store and the path is also used to key the module
on the compiler.

Currently we are not normalizing the policy path which becomes
the policy ID and also not doing so for the module path that
ends up on the compiler. This could result in the same policy
under different keys on the compiler and hence cause compile
errors of the form "multiple default rules .. found". This issue
was seen on Windows when querying OPA with a simple bundle
loaded. This fix ensures normalized paths are used as the polocy ID
for the store and module name on the compiler while using bundles.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>