topdown: Add built-in HMAC compare function
Add crypto.hmac.equal built-in function to safelly comparing hashes generated by MD5, SHA-1, SHA-256 and SHA-512 hashing algorithms.

The built-in function is a wrapped for the function Equal of package crypto/hmac.

It's useful when you need to write a policy that requires to check hash signature for request body.

Signed-off-by: Sandokan D. Arantes <sandokandias@gmail.com>