ast: Set result of groundness check on indexer's AllRules func
We perform groundness checks at index creation time and then
include result of that in the indexer's result. Previously we
did not set this value on the indexer's AllRules call as a
result of which rule evaluation for complete rules would get skipped.
This change updates the indexer's AllRules call to include the
result of the groundness checks.

Fixes: #5857

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>