[bugfix] Restore legacy ecosystem template
https://github.com/open-policy-agent/opa/pull/6029 introduced a bug
where older versions of the ecosystem page in docs were not rendered
correctly due to a missing template.

This PR restores the template referenced by the now-removed
ecosystem.html file, which still exists in older versions.

This legacy template must be retained for as long as the older docs
versions ecosystem pages need to function.

Signed-off-by: Charlie Egan <charlie@styra.com>